Added basic csv2::Writer class
